backend/requirements.txt
pandas==2.0.3
numpy==1.23.3
notebook
scikit-learn==1.3.1
matplotlib==3.7.3
python-box==6.0.2
pyYAML==6.0.1
tqdm==4.66.1
joblib==1.3.2
types-PyYAML==6.0.12.12
tensorflow==2.13.0
dvc==3.31.2
django==4.2.0
#for backend
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework-simplejwt==5.3.1
PyJWT<2.0.0
pytz==2022.6
sqlparse==0.4.3
tzdata
# tensorflow-gpu==2.4.0
python-dotenv==1.0.1
google.generativeai
